build: add macOS deployment target + portable gint64 casts
Sets Package.swift platforms to macOS 13, the minimum that satisfies
`Duration.components` and `MainActor.assumeIsolated` used in
`MainContext.swift`. Without this SwiftPM defaulted to macOS 10.13,
producing 82 availability errors on Apple toolchains.

`gint64` is `long` (== Swift `Int`) on Linux x86_64 but `long long` (==
`Int64`) on Apple arm64. `MediaStream` previously fed/returned `Int`
directly, which only happened to compile on Linux. Wrapping the three
call sites in explicit `Int(...)` / `gint64(...)` keeps the public API
unchanged and is a no-op on Linux.

+++ Sources/Adwaita/GtkWidgets/MediaStream.swift
@@ -67,12 +67,12 @@ public class MediaStream: GObjectRef {
 
     /// The current playback position in microseconds.
     public var timestamp: Int {
-        gtk_media_stream_get_timestamp(streamPointer)
+        Int(gtk_media_stream_get_timestamp(streamPointer))
     }
 
     /// The total duration in microseconds, or 0 if unknown.
     public var duration: Int {
-        gtk_media_stream_get_duration(streamPointer)
+        Int(gtk_media_stream_get_duration(streamPointer))
     }
 
     /// Whether seeking is supported.
@@ -87,7 +87,7 @@ public class MediaStream: GObjectRef {
 
     /// Seeks to the given position in microseconds.
     public func seek(_ timestamp: Int) {
-        gtk_media_stream_seek(streamPointer, timestamp)
+        gtk_media_stream_seek(streamPointer, gint64(timestamp))
     }
 
     // MARK: - Audio
